moving OntologyRelationship and OntologyRelationshipType to the model package, because they form part of the model, which is returned to the client. They are also used by OntologyGraph representation.
